Move license information to its proper locations
The BSD license goes in the relevant CMake file, and
COPYING-CMAKE-SCRIPTS is removed.  README.md gets a License section
describing the overall license of the framework, as well as a note about
the CMake scripts.

As part of this commit, I release the existing autotest images (which I
created myself) as public domain (or license them under CC0 where public
domain does not apply), ensuring they are compatible with whatever
license the package is distributed under.

REVIEW: 115423